feat(graphics): `EssenceMaterial` insertion/removal handled via observer
