+++ src/plugins/common/src/components/essence.rs
@@ -2,6 +2,7 @@ use bevy::prelude::*;
 use serde::{Deserialize, Serialize};
 
 #[derive(Component, Default, Debug, PartialEq, Clone, Copy, Serialize, Deserialize)]
+#[component(immutable)]
 pub enum Essence {
 	#[default]
 	None,
