+++ src/plugins/graphics/src/observers.rs
@@ -1,1 +1,2 @@
 pub(crate) mod insert_first_pass_camera;
+pub(crate) mod update_essence_shader;
